Update of ColloidalIce to fix some errors.
Some errors where present in the last commit. These are:
- The docstring was not indented properly
Also, the way the functions to modify a ColloidInTrap object where written, they modified the value of the colloid that called them. It is much more easy to understand if that modification has to be made explicit, and thus, the new behaviour is that the functions return a modified copy (deepcopy) of the ColloidInTrap object, that can be assigned to the ColloidalIce object.