Update to LAMMPSInterface
I fixed some things with the Reading Function. One is that the initialization of the LAMMPS Lazy read object was doing a reference of the item, and it changed the object every succesive line. This justified the inclusion of the copy library. The other is that I changed the FrameData object to be a structured array. This makes it easier to know what the columns mean. 

A future update might benefit from changing this structured object to a pandas dataframe.