On branch cleanup Your branch is up to date with 'origin/cleanup'.
Changes to be committed:
	modified:   ColloidalIce.py
	modified:   LAMMPSInterface.py
	modified:   Parameters.py
	deleted:    ReadColloidalIceSimulation.py
	modified:   Vertices.py
	modified:   ../IceNumericsUsage.ipynb

Detailed change description

These are the changes made to the ColloidalIce package
ColloidInTrap is now colloid_in_trap

I changed the way the trap separation is specified. In the new version trap_sep is an attribute of the TrapGeometry class which determines the separation between traps.

The colloid position is no longer specified as an argument to the initialization of the colloid_in_trap class. Now the colloid is automatically initialized in the center of the trap specified by the direction attribute. It is still an attribute of colloid.

The colloid_in_trap attributes, center, direction and colloid are no longer Vector objects. This is with the intention of geting rid of the Vector class which doesn't work very well. Now they are numpy arrays of 3 elements.
The direction attribute's magnitude shouldn't matter. In principle, the direction vector is unitary, but we should ensure that when it is used its always divided by it's magnitude before, in case it is changed outside the class methods.

The display method for colloid_in_trap now uses the trap_sep attribute to draw islands of the proper length. I also set the patch fill color again, because if particles are below the trap patches it means that the match from the simulation to the colloidal_ice object is not correct. Setting them to transparent is simply a way to fix the symptom but not the problem.

In the flip and bias functions I only removed the option to have Vector objects

ColloidalIce is now colloidal_ice

colloidal_ice now inherits from list, instead of dict, which makes access to the colloid_in_trap objects easier.
When initialized, colloidal_ice becomes a list of colloid_in_trap objects with some extra parameters.

The lattice parameter is no longer a parameter of the colloidal_ice object. This has disadvantages which I need to figure out, but on the other hand, it really shouldn't be here. In any case, it should be a part of "worldparameters" but probably not even that.
The original reason for the lattice parameter to be here is that it was used to calcluate things like the region's extra space and the dipole interaction cutoff length. The cutoff frequency is now a parameter of the simulation, and the region's extra space should be either calculated from the region span, from the particles size or specified as a parameter. Right now it's simply padded with a constant length, but this should be fixed in the future.

The ordering parameter also has no place on the colloidal_ice class. This prevents us to have some colloids which start random and some which start at predefined positions. Instead I included an attribute "initial_position" in the TrapGeometry class (which can be set by the "InitialPosition" keyword argument) and can have the values "fixed" or "random".
In the intialization of a colloidal_ice object, the initial_position is set by the ordering property of the Spin system which is used to define it. We should think if it makes sense to leave it this way or if there is a more natural way of determining it.

The display functions of most objects now accept as optional argument the axis only. Before they accepted a general DisplayObject which could also be a figure. I think this is cumbersome and unnecessary.

FrameDataToColloids is now framedata_to_colloids.
There was an error in the way the run argument was handled. run should specify the number of run we extract from the framedata object, whith the python standard indexing (0 indexing). That means that run==number_of_runs will give the traps position, which is wrong. Now I check the parameter for (run)<=number_of_runs. Otherwise I raise an error.

The matching between the colloids in the framedata object and the colloidal_ice object is done more efficiently by a cKDTree. This yields a directory where_in_frame with the property that the element where_in_frame[i] is the location of the framedata row corresponding to  colloidal_ice[i]

CalculateEnergy is now calculate_energy, and it works the same as before

These are the changes made to the Parameters package
I deleted the attribute TrapGeometry.trap_sep_ratio
I added the attribute TrapGeometry.trap_sep
I added the attribute TrapGeometry.initial_position

The definition of the set_region method for WorldParameter now calculates the region using a numpy array.
Also, the region padding is now a constant value. We should change this in the future.

I added the attribute SimulationParameters.dipole_cutoff

ReadColloidalIceSimulation.py was deleted

Vertices.py
Here I only changed small things to comply with the new standard.