Fixed bug in ColloidalIce.FrameDataToColloids
Fixed a bug in FrameDataToColloids which caused some colloids to be connected to simulation points that didn't correspond. This was the reason that sometimes, in the display function, the traps were drawn over the colloids. Since the FrameDataToColloids is fixed, the display function should always display the traps before the colloids, and therefore I am filling the patches again.

I also throw an error now when a nonexistent run is requested.