Improved HoneycombSpinIceGeometry for efficiency
The creation of honeycomb spin ice is done by creating six spins per cell and then eliminating repetitions. In the previous version, the search for repetitions was hugely inefficient. I incorporated a cKDTree to look for neighbors within a tolerance and this decreased the time taken to create large systems by several orders of magnitude.